增加欧式距离

+++ zing-biz/src/main/java/com/yibao/biz/bizenun/QuestionsEnum.java
@@ -12,23 +12,23 @@ public enum QuestionsEnum {
      * 问题模板索引
      */
 
-    DISEASE_PREVENT(0.0, "ill 预防", "ill/prevent.txt"),
-    DISEASE_CAUSE(1.0, "ill 诱因", "ill/cause.txt"),
-    DISEASE_DESC(2.0, "ill 简介", "ill/desc.txt"),
-    SYMPTOM_CHECK(3.0, "sym 检查", "sym/check.txt"),
-    DISEASE_CUREWAY(4.0, "ill 治疗方式", "ill/cureway.txt"),
-    DISEASE_CHECK(6.0, "ill 诊断检查", "ill/check.txt"),
-    DISEASE_DRUG(7.0, "ill 常用药品", "ill/drug.txt"),
-    DISEASE_SYMPTOM(8.0, "ill 症状", "ill/sym.txt"),
-    DISEASE_DISEASE(9.0, "ill 并发症", "ill/dis.txt"),
-    DISEASE_DEPARTMENT(10.0, "ill 所属科室", "ill/department.txt"),
-    SYMPTOM_DISEASE(11.0, "sym 导致疾病", "sym/dis.txt"),
-    SYMPTOM_DRUG(5.0, "sym 用药", "sym/drug.txt"),
-    SYMPTOM_CUREWAY(12.0, "sym 用药", "sym/cureway.txt"),
+    DISEASE_PREVENT(0, "ill 预防", "ill/prevent.txt"),
+    DISEASE_CAUSE(1, "ill 诱因", "ill/cause.txt"),
+    DISEASE_DESC(2, "ill 简介", "ill/desc.txt"),
+    SYMPTOM_CHECK(3, "sym 检查", "sym/check.txt"),
+    DISEASE_CUREWAY(4, "ill 治疗方式", "ill/cureway.txt"),
+    DISEASE_CHECK(6, "ill 诊断检查", "ill/check.txt"),
+    DISEASE_DRUG(7, "ill 常用药品", "ill/drug.txt"),
+    DISEASE_SYMPTOM(8, "ill 症状", "ill/sym.txt"),
+    DISEASE_DISEASE(9, "ill 并发症", "ill/dis.txt"),
+    DISEASE_DEPARTMENT(10, "ill 所属科室", "ill/department.txt"),
+    SYMPTOM_DISEASE(11, "sym 导致疾病", "sym/dis.txt"),
+    SYMPTOM_DRUG(5, "sym 用药", "sym/drug.txt"),
+    SYMPTOM_CUREWAY(12, "sym 用药", "sym/cureway.txt"),
     ;
 
     @Getter
-    private Double index;
+    private Integer index;
 
     @Getter
     private String question;
@@ -36,7 +36,7 @@ public enum QuestionsEnum {
     @Getter
     private String filePath;
 
-    QuestionsEnum(Double index, String question, String filePath) {
+    QuestionsEnum(Integer index, String question, String filePath) {
         this.index = index;
         this.question = question;
         this.filePath = filePath;
